% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gg_colors.R
\name{gg_colors}
\alias{gg_colors}
\title{Generate default ggplot2 color palette}
\usage{
gg_colors(n)
}
\arguments{
\item{n}{The number of colors to generate}
}
\value{
A vector of colors
}
\description{
Generate default ggplot2 color palette
}
\examples{
gg_colors(4) #First four ggplot2 colors
}

