% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/musicnot.R
\name{musicnot}
\alias{musicnot}
\title{Generate music notation graph of social interactions}
\usage{
musicnot(df, ranks = NULL, arrowsize = 0.3, arrowweight = 0.75,
  pointsize = 3, ylabel = NULL, colors = "default", gridcolors = NULL,
  gridlinesize = 1, labels = "rank", behav = "default")
}
\arguments{
\item{df}{A dataframe with at least 3 columns in the order: time, id1, id2.
Every observation/row should be a mututally exclusive behavioral event.
The first variable, time, can be in numeric or date/time format. The 2nd and
3rd variables, id1 and id2 should be character variables denoting the id of
the individual (id1) directing each behavior to the other individual (id2) that
receives the behavior.  If want to discriminate lines by different behaviors,
a 4th column should be the behavior of interest.  Other columns are permissible
in the dataframe that come after these required variables.}

\item{ranks}{The default is NULL meaning that individuals will be plotted
in alphanumerical order.  Otherwise, a character vector indicating the rank
of individuals (from highest to lowest) which will determine the plotted order
of individuals.}

\item{arrowsize}{The size of the arrowhead.}

\item{arrowweight}{The weight of the arrowhead.}

\item{pointsize}{The size of points.}

\item{ylabel}{The y-axis title.}

\item{colors}{The colors to be used for plotting.  The defualt is
\strong{\code{type}="default"} meaning the default ggplot2 color palette is
used.  Other options are:  \strong{\code{type}="terrain"}, \strong{\code{type}="heat"}
\strong{\code{type}="cm"}, \strong{\code{type}="random"}, \strong{\code{type}="grays"}
\strong{\code{type}="rainbow"}, \strong{\code{type}="topo"}, or alternatively a
character vector of color names or codes of can be used.}

\item{gridcolors}{default is to use grey horizontal grid lines,
but if \strong{\code{type}="gridcolors=T"} then horizontal grid lines of the same
color as each individual will be used.}

\item{gridlinesize}{Weight/size of gridlines.}

\item{labels}{Default is \strong{\code{type}="rank"} which will use numeric values as
y-axis labels, but if \strong{\code{type}="name"} then the individual's id will be used.}

\item{behav}{Default is \strong{\code{type}="default"} which is to not differentiate
different behaviors in the plot, but if \strong{\code{type}="yes"} then lines will be
differentially styled (dotted, dashed, etc.) to distinguish different behaviors.  A
character vector with the names of each behavior can be used to determine the order in
which different line styles are to be applied.  To change from \strong{\code{type}="default"}
requires a 4th variable in the dataframe identifying the behavior at each observation.}
}
\value{
A music notation graph of social interactions.
}
\description{
Generate music notation graph of social interactions
}
\section{References}{

Ivan D. Chase, (2006),
Music notation: a new method for visualizing social interaction in animals and humans
Frontiers in Zoology 3: 18
}

\section{Further details}{

tba
}
\examples{
musicnot(flies)
musicnot(flies, gridcolors=T)
musicnot(flies, gridcolors=T, gridlinesize=0.5)
musicnot(flies, gridcolors=T, gridlinesize=0.5, colors="topo")
musicnot(flies, gridcolors=T, gridlinesize=0.5, colors="rainbow")
musicnot(flies, gridcolors=T, gridlinesize=0.5, labels="name")
musicnot(flies, gridcolors=T, gridlinesize=0.5, behav="yes")
}

