% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/random_datetime.R
\name{random_datetime}
\alias{random_datetime}
\title{Generate random dates and times between two dates}
\usage{
random_datetime(N, st = "2015/01/01", et = "2015/12/31")
}
\arguments{
\item{N}{the number of random dates/times to produce}

\item{st}{the start date in the format e.g. "2014/11/17"}

\item{et}{the end date in the format e.g. "2014/11/17"}
}
\value{
A sequence of random dates and times
}
\description{
Generate random dates and times between two dates
}
\examples{
random_datetime(100, st="2015/01/01", et="2015/01/31")
}

